+++ ex00/BitcoinExchange.hpp
@@ -6,7 +6,7 @@
 /*   By: gdaignea <gdaignea@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/24 15:04:59 by gdaignea          #+#    #+#             */
-/*   Updated: 2024/10/28 16:19:00 by gdaignea         ###   ########.fr       */
+/*   Updated: 2024/10/28 17:00:04 by gdaignea         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ class Bitcoin {
 
 	private:
 
-	std::map<std::string, double>	_ratesData;
+	std::map<std::string, float>	_ratesData;
 
 	public:
 
@@ -33,7 +33,7 @@ class Bitcoin {
 	~Bitcoin();
 
 	Bitcoin&	operator=(Bitcoin const&);
-	std::map<std::string, double> const&	getRatesData() const;
+	std::map<std::string, float> const&	getRatesData() const;
 
 	void	extractData(std::ifstream& file);
 	void	displayBtcAmount(std::ifstream& inputFile);
